% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_poster.R
\name{upload_poster}
\alias{upload_poster}
\title{Upload posters from the mediux files}
\usage{
upload_poster(file)
}
\arguments{
\item{file}{The .yaml file with the MediUX posters.}
}
\value{
Verbatim for each poster if it uploaded or not
}
\description{
Upload posters from the mediux files
}
\examples{
\dontrun{
upload_poster("test.yaml")
}
}
