% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_poster_movies.R
\name{upload_poster_movies}
\alias{upload_poster_movies}
\title{Generic function to upload the movies posters}
\usage{
upload_poster_movies(input)
}
\arguments{
\item{input}{A data.frame of movies and poster url}
}
\value{
The verbatim
}
\description{
Generic function to upload the movies posters
}
