% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_poster_series.R
\name{upload_poster_series}
\alias{upload_poster_series}
\title{Upload posters for TV series from MediUX YAML files}
\usage{
upload_poster_series(input)
}
\arguments{
\item{input}{The `.yaml` file containing the MediUX shows as a `list` format.}
}
\value{
Prints a message for each series poster indicating whether it was successfully uploaded.
}
\description{
Upload posters for TV series from MediUX YAML files
}
\keyword{internal}
