% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_poster.R
\name{upload_poster}
\alias{upload_poster}
\title{Upload posters from MediUX YAML file}
\usage{
upload_poster(file, section = "all")
}
\arguments{
\item{file}{The `.yaml` file containing the MediUX poster definitions.}

\item{section}{Which section of the file you want to update.
By default, all sections are processed. Accepts `"movies"`, `"shows"`, or `"all"`.}
}
\value{
Prints a message for each poster indicating whether it was successfully uploaded.
}
\description{
Upload posters from MediUX YAML file
}
\examples{
\dontrun{
upload_poster("test.yaml")
}
}
