% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_poster_movies.R
\name{upload_poster_movies}
\alias{upload_poster_movies}
\title{Upload posters for movies from MediUX YAML files}
\usage{
upload_poster_movies(input)
}
\arguments{
\item{input}{The `data.frame` of movies.}
}
\value{
Prints a message for each movie poster indicating whether it was successfully uploaded.
}
\description{
Upload posters for movies from MediUX YAML files
}
\keyword{internal}
